refactor: Follow file naming conventions (#201)
* add ms-identity-web to requirements

* fix:removed expired and disposed items in allocate asset

* search employee issue on first letter is solved

* feat: deallocate asset front end completed without api, backend started

* feat: deallocation front end and backend completed

* fix: Rename import path in unassign_service

* fix: Remove bug in unassign_service

* refactor: Follow file naming conventions

---------
