+++ exam_django/asset/views/__init__.py
@@ -13,7 +13,7 @@ from .data_import_view import DataImportView
 from .QueryBuilderView import QueryBuilderView
 from .asset_export_view import AssetExportView
 from .asset_lifecycle_view import AssetLifeCycleView
-from .unassign_asset_view import UnassignAssetView
+from .asset_unassign_view import UnassignAssetView
 
 __all__ = [
     "AssetView",
@@ -25,12 +25,12 @@ __all__ = [
     "MemoryView",
     "EmployeeView",
     "AssignAssetView",
+    "UnassignAssetView",
     "AssetSearchWithFilterView",
     "AssetLogView",
     "DataImportView",
     "AssetExportView",
     "UserAgentAssetView",
     "AssetLifeCycleView",
-    "QueryBuilderView"
-    "UnassignAssetView",
+    "QueryBuilderView",
 ]
